requirements.txt
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
autocommand==2.2.2
babel==2.17.0
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dbus-python==1.4.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
docstring-to-markdown==0.17
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.0
fqdn==1.5.1
gps==3.27.2
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
i3ipc==2.2.1
idna==3.11
importlib_metadata==8.7.1
inflect==7.5.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lensfun==0.3.4
lsprotocol==2025.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
meson==1.10.0
mistune==3.2.0
more-itertools==10.8.0
mplsoccer==1.6.1
msgpack==1.1.2
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
nftables==0.1
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.4.0
nwg-displays==0.3.26
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
perf==0.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
pooch==1.8.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.29.0
pycparser==2.23
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
pynvim==0.6.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-linux-procfs==0.7.4
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.3.0
python-lsp-server==1.14.0
python-xlib==0.33
pytokens==0.3.0
pytz==2025.2
pyudev==0.24.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
Reflector==2023.6.28.0.36.1
requests==2.32.5
requests-cache==1.2.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruff==0.14.10
scipy==1.16.3
seaborn==0.13.2
Send2Trash==2.0.0
setproctitle==1.3.7
setuptools==80.9.0
shtab==1.8.0
six==1.17.0
soupsieve==2.8.1
stack-data==0.6.3
statsbombpy==1.16.0
termcolor==3.2.0
terminado==0.18.1
tinycss2==1.4.0
tldr==3.4.3
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
ufw==0.36.2
ujson==5.11.0
uri-template==1.3.0
url-normalize==2.2.1
urllib3==2.6.0
validate==5.0.9
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.45.1
widgetsnbextension==4.0.15
wxPython==4.2.3
zipp==3.23.0
zstandard==0.25.0
